+++ packages/js/plugins/ipfs/src/resolvers.ts
@@ -1,22 +1,37 @@
 import { IpfsPlugin } from "./";
-import { ResolveResult, Query, Mutation } from "./w3";
+import { ResolveResult, Options, Query, Mutation, QueryEnv } from "./w3";
 
-export const mutation = (ipfs: IpfsPlugin): Mutation.Module => ({
-  addFile: async (input: Mutation.Input_addFile) => {
-    const { hash } = await ipfs.add(input.data);
-    return hash.toString();
-  },
-});
+const getOptions = (
+  input: Options | undefined | null,
+  env: QueryEnv
+): Options => {
+  const options = input || {};
+
+  if (
+    options.disableParallelRequests === undefined ||
+    options.disableParallelRequests === null
+  ) {
+    options.disableParallelRequests = env.disableParallelRequests;
+  }
+
+  return options;
+};
 
 export const query = (ipfs: IpfsPlugin): Query.Module => ({
   catFile: async (input: Query.Input_catFile) => {
-    return await ipfs.cat(input.cid, input.options || undefined);
+    const queryEnv = ipfs.getEnv("query") as QueryEnv;
+    const options = getOptions(input.options, queryEnv);
+    return await ipfs.cat(input.cid, options);
   },
   resolve: async (input: Query.Input_resolve): Promise<ResolveResult> => {
-    return await ipfs.resolve(input.cid, input.options || undefined);
+    const queryEnv = ipfs.getEnv("query") as QueryEnv;
+    const options = getOptions(input.options, queryEnv);
+    return await ipfs.resolve(input.cid, options);
   },
   // uri-resolver.core.web3api.eth
   tryResolveUri: async (input: Query.Input_tryResolveUri) => {
+    const queryEnv = ipfs.getEnv("query") as QueryEnv;
+
     if (input.authority !== "ipfs") {
       return null;
     }
@@ -37,8 +52,10 @@ export const query = (ipfs: IpfsPlugin): Query.Module => ({
     for (const manifestSearchPattern of manifestSearchPatterns) {
       try {
         manifest = await ipfs.catToString(
-          `${input.path}/${manifestSearchPattern}`,
-          { timeout: 5000 }
+          `${input.path}/${manifestSearchPattern}`, {
+            timeout: 5000,
+            disableParallelRequests: queryEnv.disableParallelRequests,
+          }
         );
       } catch (e) {
         // TODO: logging
@@ -54,16 +71,28 @@ export const query = (ipfs: IpfsPlugin): Query.Module => ({
     }
   },
   getFile: async (input: Query.Input_getFile) => {
+    const queryEnv = ipfs.getEnv("query") as QueryEnv;
+
     try {
       const { cid, provider } = await ipfs.resolve(input.path, {
         timeout: 5000,
+        disableParallelRequests: queryEnv.disableParallelRequests,
       });
 
       return await ipfs.cat(cid, {
-        provider: provider,
+        provider,
+        timeout: 20000,
+        disableParallelRequests: true,
       });
     } catch (e) {
       return null;
     }
+  },
+});
+
+export const mutation = (ipfs: IpfsPlugin): Mutation.Module => ({
+  addFile: async (input: Mutation.Input_addFile) => {
+    const { hash } = await ipfs.add(input.data);
+    return hash.toString();
   },
 });