+++ packages/schema/parse/src/transform/addFirstLast.ts
@@ -22,9 +22,9 @@ export const addFirstLast: TypeInfoTransforms = {
     TypeInfo: (typeInfo: TypeInfo): TypeInfo => ({
       ...typeInfo,
       objectTypes: setFirstLast(typeInfo.objectTypes),
-      queryTypes: setFirstLast(typeInfo.queryTypes),
+      moduleTypes: setFirstLast(typeInfo.moduleTypes),
       importedObjectTypes: setFirstLast(typeInfo.importedObjectTypes),
-      importedQueryTypes: setFirstLast(typeInfo.importedQueryTypes),
+      importedModuleTypes: setFirstLast(typeInfo.importedModuleTypes),
     }),
   },
 };