+++ packages/test-cases/cases/compose/004-empty-types/01-inherited/output/mutation.ts
@@ -1,6 +1,6 @@
 import {
   createMethodDefinition,
-  createQueryDefinition,
+  createModuleDefinition,
   createObjectDefinition,
   createObjectPropertyDefinition,
   createInterfaceImplementedDefinition,
@@ -31,9 +31,9 @@ export const typeInfo: TypeInfo = {
       ],
     },
   ],
-  queryTypes: [
+  moduleTypes: [
     {
-      ...createQueryDefinition({ type: "Mutation" }),
+      ...createModuleDefinition({ type: "Mutation" }),
       imports: [],
       interfaces: [],
       methods: [